requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.15.2
Pillow==10.4.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
boto3==1.34.162
django-storages==1.14.4
python-decouple==3.8
drf-spectacular==0.27.2
